requirements.txt
# Core dependencies
numpy>=1.24.0
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.2.0

# PDF parsing (install all for best coverage)
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pymupdf>=1.23.0

# OpenAI LLM
openai>=1.0.0
python-dotenv>=1.0.0

# Optional: better sentence splitting
nltk>=3.8.0


